pyproject.toml
[project]
name = "decision-games"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.0.0",
    "ruff>=0.14.10",
]

[tool.uv.workspace]
members = [
    "poker_ui",
]
